ai_backend/app/requirements.txt
fastapi
uvicorn[gunicorn]
openai
pydantic
motor
python-dotenv
requests
